add lib sweetAlert no login

+++ scripts/login.js
@@ -15,7 +15,11 @@ function checkIfAnyRoleIsChecked(){
 function cadastrar() {
     // Checa se alguma role foi checada.
     if(checkIfAnyRoleIsChecked() === false){
-        alert('Marque alguma Role!');
+        Swal.fire(
+            'Algo deu errado...',
+            'Marque alguma role!',
+            'error'
+        );
         return;
     }
     // Inicia a massa de dados (payload)
@@ -37,13 +41,26 @@ function cadastrar() {
     })
     .then(response => response.json())
     .then(response => {
-        alert('Cadastrado com sucesso!');
+        Swal.fire(
+            'Bom trabalho!',
+            'Cadastrado com sucesso!',
+            'success'
+        );
 
-        localStorage.setItem("userName", response.fullName);
-        localStorage.setItem("role", response.role === "dev" ? "Desenvolvedor" : "Cliente");
-        localStorage.setItem("idClient", response.id);
-
-        window.location.href = "list.html";
+        Swal.fire({
+            title: 'Bom trabalho!',
+            text: "Cadastrado com sucesso!",
+            icon: 'success',
+            confirmButtonText: 'Ok!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                localStorage.setItem("userName", response.fullName);
+                localStorage.setItem("role", response.role === "dev" ? "Desenvolvedor" : "Cliente");
+                localStorage.setItem("idClient", response.id);
+        
+                window.location.href = "list.html";    
+            }
+          })
     })
     
 }
